+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/lib/auth/authOptions";
+import { authOptions } from "@/lib/auth/options";
 import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
