+++ src/app/auth/login/page.tsx
@@ -1,5 +1,5 @@
 import LoginForm from "@/app/_modules/auth/components/LoginForm"
-import { authOptions } from "@/lib/auth/authOptions"
+import { authOptions } from "@/lib/auth/options"
 import { getServerSession } from "next-auth"
 import { redirect } from "next/navigation";
