+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
 import LogoutButton from "./_modules/auth/components/LogoutButton";
-import { authOptions } from "@/lib/auth/authOptions";
+import { authOptions } from "@/lib/auth/options";
 
 const HomePage = async () => {
   const session = await getServerSession(authOptions);
