+++ src/lib/auth/options.ts
